test : 학습테스트 작성
